// src/app/phonePal/model/member-dialog.model.ts
export class MemberDialog {
  // Anthem Programs
  vMedicationIssues: string;
  vPCP: string;
  vStarHeidis: string;
  vCriticalCare: string;
  vMemberServices: string;
  vBehaviorHealth: string;
  vMyAdvocate: string;
  vOtherPrograms: string;

  // Barriers
  vFinance: string;
  vDepression: string;
  vGrief: string;
  vPain: string;
  vTransportation: string;
  vOtherBarriers: string;

  // Care coordination
  vCarePCP: string;
  vCaseManager: string;
  vContactedReferral: string;
  vTransferredMember: string;
  vOtherCare: string;

  // Community Resources
  vCommunityTransportation: string;
  vSeniorsCenter: string;
  vHomeMaintenance: string;
  vSupportGroups: string;
  vCommunityNo: string;
  vMealsOnWheels: string;
  vSeniorCenterWithoutWalls: string;
  vHousing: string;
  vCommunityYes: string;
  vOtherCommunity: string;

  // Physical Activity
  vGym: string;
  vJogging: string;
  vNiftyAfterFifty: string;
  vWalking: string;
  vOtherActivities: string;    
}